Use uint64 instead of uint

+++ ts/run/win.go
@@ -1,40 +1,40 @@
 package run
 
 type RunSum struct {
-	win, idx, cnt uint
+	win, idx, cnt uint64
 	vals          []float64
 	sum           float64
 }
 
 type RunMax struct {
-	win, idx, cnt uint
+	win, idx, cnt uint64
 	vals          []float64
 }
 
 type RunMin struct {
-	win, idx, cnt uint
+	win, idx, cnt uint64
 	vals          []float64
 }
 
 type RunAvg struct {
-	win, idx, cnt uint
+	win, idx, cnt uint64
 	vals          []float64
 	sum           float64
 }
 
 type RunFst struct {
-	win, idx, cnt uint
+	win, idx, cnt uint64
 	vals          []float64
 }
 
 // Simple Linear Regression over timeline
 type RunSlr struct {
-	win, idx, cnt uint
+	win, idx, cnt uint64
 	vals, ws      []float64
 	varX, sumY    float64
 }
 
-func NewRunSum(win uint) *RunSum {
+func NewRunSum(win uint64) *RunSum {
 	return &RunSum{win: win, vals: make([]float64, win)}
 }
 
@@ -48,7 +48,7 @@ func (run *RunSum) App(val float64) (sum float64) {
 	return run.sum
 }
 
-func NewRunMax(win uint) *RunMax {
+func NewRunMax(win uint64) *RunMax {
 	return &RunMax{win: win, vals: make([]float64, win)}
 }
 
@@ -58,7 +58,7 @@ func (run *RunMax) App(val float64) (max float64) {
 		run.cnt++
 	}
 	max = run.vals[0]
-	for i := uint(1); i < run.cnt; i++ {
+	for i := uint64(1); i < run.cnt; i++ {
 		if run.vals[i] > max {
 			max = run.vals[i]
 		}
@@ -67,7 +67,7 @@ func (run *RunMax) App(val float64) (max float64) {
 	return
 }
 
-func NewRunMin(win uint) *RunMin {
+func NewRunMin(win uint64) *RunMin {
 	return &RunMin{win: win, vals: make([]float64, win)}
 }
 
@@ -77,7 +77,7 @@ func (run *RunMin) App(val float64) (min float64) {
 		run.cnt++
 	}
 	min = run.vals[0]
-	for i := uint(1); i < run.cnt; i++ {
+	for i := uint64(1); i < run.cnt; i++ {
 		if run.vals[i] < min {
 			min = run.vals[i]
 		}
@@ -86,7 +86,7 @@ func (run *RunMin) App(val float64) (min float64) {
 	return
 }
 
-func NewRunAvg(win uint) *RunAvg {
+func NewRunAvg(win uint64) *RunAvg {
 	return &RunAvg{win: win, vals: make([]float64, win)}
 }
 
@@ -100,7 +100,7 @@ func (run *RunAvg) App(val float64) (avg float64) {
 	return run.sum / float64(run.cnt)
 }
 
-func NewRunFst(win uint) *RunFst {
+func NewRunFst(win uint64) *RunFst {
 	return &RunFst{win: win, vals: make([]float64, win)}
 }
 
@@ -117,10 +117,10 @@ func (run *RunFst) App(val float64) (fst float64) {
 	}
 }
 
-func NewRunSlr(win uint) *RunSlr {
+func NewRunSlr(win uint64) *RunSlr {
 	run := &RunSlr{win: win, vals: make([]float64, win), ws: make([]float64, win)}
 	avg := float64(win-1) / float64(2)
-	for i := uint(0); i < win; i++ {
+	for i := uint64(0); i < win; i++ {
 		run.ws[i] = float64(i) - avg
 		run.varX += run.ws[i] * run.ws[i]
 	}
@@ -136,12 +136,12 @@ func (run *RunSlr) App(val float64) (slope float64) {
 		return
 	}
 	cov, yAvg := float64(0), run.sumY/float64(run.win)
-	for i := uint(0); i < run.win; i++ {
+	for i := uint64(0); i < run.win; i++ {
 		cov += run.ws[i] * (run.vals[cnt2idx(run.win, run.idx+i)] - yAvg)
 	}
 	return cov / run.varX
 }
 
-func cnt2idx(win, cnt uint) uint {
+func cnt2idx(win, cnt uint64) uint64 {
 	return cnt % win
 }
